+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 
 ### Added
-- Support for `group` (`grouping`) and `having` (`having_grouped`).
+- Support for `group` (`grouping`) and `having` (`when_having`).
 - Support for sifters.
 - More descriptive error messages when a column or association is not found.